peg-0.1.20 (r113)
Experimental support for variables and semantic values in ‘inline actions’ executed
during recognition (separate from the parser actions that are run after the input is fully recognised).

Don't print `#line` directives for headers if `-P` was given and
pathnames on WIN32 have `\` changed to `/` to keep `#line` directives happy
(thanks to Ben Hines for the problem report).